WIP: Clone a slide, add 2 text boxes by default when adding a slide
Clone a slide works well for images, doesn't work for text boxes, there is an issue that the text box is considered "Proxy" once it has text in it. Should be fixed immediately.
Adding two text boxes, needs positioning and sizing of the boxes. Right now this happens with 'px', instead a way should be found in order for percentages to be stored and used.